Added in all of the data into my database tables, Added in more code into each controller which is used for the Create, edit and delete Methods.
Added in Regions to make the code slick and clean.

Added in another data processor into my buisness logic this is for the sale controller in the main program. Updated the Customer Data page so it doesnt sow confirm email.

Also Updated the Product Data page which has the sale ID before the Product ID, these two updates are more of just a UserInterface Change rather than back end code.